-added a service interface and its implementation; -added javax.ws.rs to pom;

+++ book-storage-REST/src/com/bookstorage/model/Book.java
@@ -1,22 +1,22 @@
 package com.bookstorage.model;
 
 public class Book {
-	private long id;
+	private int id;
 	private String title;
 	private String author;
 	private int location_id;
 	
-	public Book(long id, String title, String author, int location_id) {
+	public Book(int id, String title, String author, int location_id) {
 		this.id = id;
 		this.title = title;
 		this.author = author;
 		this.location_id = location_id;
 	}
 	
-	public long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getTitle() {
